FOLSPRINGB-126: Reject empty and null system user password
The folio-spring-system-user library should reject an empty or null system user password.

Use case:

As a sysop I may accidentially forget to set the environment variable.

For example forgetting to add SYSTEM_USER_PASSWORD, or forget to add the secret to a deployment in Rancher, or having a typo in the variable name like SYSTEM_USERS_PASSWORD.

When the system user password is not set or is empty the attempt to create the system user when enabling the module for a tenant (or upgrading the module for a tenant) should fail with an error message about the missing password.

The module MUST NOT set an empty password.

This is security by default and is required for GDPR compliance (GDPR Article 25 "Data protection by design and by default").

Therefore folio-spring-system-user MUST reject a system user with null or empty password.