FOLSPRINGS-180: Token expiration off by 1 minute in test
https://folio-org.atlassian.net/browse/FOLSPRINGS-180

Purpose

Four tests in SystemUserServiceTest fail with
```
org.opentest4j.AssertionFailedError:
Expecting:
 <2024-12-06T23:30:00Z>
to be equal to:
 <2024-12-06T23:29:00Z>
but was not.
```

when the test of that class starts at second :59 and executes during the next minute at second :00.

TokenUtils.calculateTokenExpirationForUser makes useless truncation of the ttl.

Tasks:
* Change test setup to work at any second.
* Remove truncation from TokenUtils.calculateTokenExpirationForUser ttl calculation.

Approach

In SystemUserServiceTest call Instant.now() and reuse it for all values in all tests, this allows for exact comparison.
In SystemUserServiceTest use .isCloseTo(..., within(10, ChronoUnit.SECONDS)) when the actual time value has been generated by TokenUtils and depends on the execution speed.
In TokenUtils.calculateTokenExpirationForUser use Duration#dividedBy(2).